refactor model package and update DG accordingly.
