+++ src/main/java/clzzz/helper/logic/Logic.java
@@ -7,7 +7,7 @@ import clzzz.helper.logic.commands.CommandResult;
 import clzzz.helper.logic.commands.exceptions.CommandException;
 import clzzz.helper.logic.parser.exceptions.ParseException;
 import clzzz.helper.model.ReadOnlyPetTracker;
-import clzzz.helper.model.pet.FoodCollection;
+import clzzz.helper.model.foodcollection.FoodCollection;
 import clzzz.helper.model.pet.Pet;
 import clzzz.helper.model.slot.Slot;
 import clzzz.helper.ui.DisplaySystemType;
