+++ src/main/java/clzzz/helper/logic/commands/pet/EditPetCommand.java
@@ -28,7 +28,7 @@ import clzzz.helper.model.pet.Gender;
 import clzzz.helper.model.pet.Name;
 import clzzz.helper.model.pet.Pet;
 import clzzz.helper.model.pet.Species;
-import clzzz.helper.model.tag.Tag;
+import clzzz.helper.model.pet.Tag;
 
 /**
  * Edits the details of an existing pet in the pet tracker.
