+++ src/main/java/clzzz/helper/logic/parser/ParserUtil.java
@@ -15,7 +15,7 @@ import clzzz.helper.model.pet.Food;
 import clzzz.helper.model.pet.Gender;
 import clzzz.helper.model.pet.Name;
 import clzzz.helper.model.pet.Species;
-import clzzz.helper.model.tag.Tag;
+import clzzz.helper.model.pet.Tag;
 import clzzz.helper.ui.DisplaySystemType;
 
 /**
