+++ src/main/java/clzzz/helper/logic/parser/pet/AddPetCommandParser.java
@@ -25,7 +25,7 @@ import clzzz.helper.model.pet.Gender;
 import clzzz.helper.model.pet.Name;
 import clzzz.helper.model.pet.Pet;
 import clzzz.helper.model.pet.Species;
-import clzzz.helper.model.tag.Tag;
+import clzzz.helper.model.pet.Tag;
 
 /**
  * Parses input arguments and creates a new AddPetCommand object.
