+++ src/main/java/clzzz/helper/model/foodcollection/FoodAmountAndPet.java
@@ -1,10 +1,11 @@
-package clzzz.helper.model.pet;
+package clzzz.helper.model.foodcollection;
 
 import static clzzz.helper.commons.util.AppUtil.checkArgument;
 import static java.util.Objects.requireNonNull;
 
 import java.util.Objects;
 
+import clzzz.helper.model.pet.Pet;
 import clzzz.helper.ui.DisplaySystemType;
 import clzzz.helper.ui.list.DisplayItem;
