+++ src/main/java/clzzz/helper/model/pet/Pet.java
@@ -7,7 +7,6 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
-import clzzz.helper.model.tag.Tag;
 import clzzz.helper.ui.DisplaySystemType;
 import clzzz.helper.ui.list.DisplayItem;
