+++ src/main/java/clzzz/helper/model/util/SampleDataUtil.java
@@ -17,8 +17,8 @@ import clzzz.helper.model.pet.Gender;
 import clzzz.helper.model.pet.Name;
 import clzzz.helper.model.pet.Pet;
 import clzzz.helper.model.pet.Species;
+import clzzz.helper.model.pet.Tag;
 import clzzz.helper.model.slot.Slot;
-import clzzz.helper.model.tag.Tag;
 
 
 /**
