+++ src/main/java/clzzz/helper/ui/OverallStats.java
@@ -9,7 +9,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import clzzz.helper.commons.util.DateTimeUtil;
-import clzzz.helper.model.pet.FoodCollection;
+import clzzz.helper.model.foodcollection.FoodCollection;
 import clzzz.helper.model.pet.Pet;
 import clzzz.helper.model.pet.Species;
 import clzzz.helper.model.slot.Slot;
