+++ src/test/java/clzzz/helper/logic/parser/pet/AddPetCommandParserTest.java
@@ -43,7 +43,7 @@ import clzzz.helper.model.pet.Gender;
 import clzzz.helper.model.pet.Name;
 import clzzz.helper.model.pet.Pet;
 import clzzz.helper.model.pet.Species;
-import clzzz.helper.model.tag.Tag;
+import clzzz.helper.model.pet.Tag;
 import clzzz.helper.testutil.pet.PetBuilder;
 
 public class AddPetCommandParserTest {
