+++ src/test/java/clzzz/helper/logic/parser/pet/EditPetCommandParserTest.java
@@ -43,7 +43,7 @@ import clzzz.helper.model.pet.DateOfBirth;
 import clzzz.helper.model.pet.Gender;
 import clzzz.helper.model.pet.Name;
 import clzzz.helper.model.pet.Species;
-import clzzz.helper.model.tag.Tag;
+import clzzz.helper.model.pet.Tag;
 import clzzz.helper.testutil.pet.EditPetDescriptorBuilder;
 
 public class EditPetCommandParserTest {
